requirements.txt
fastapi
uvicorn
pydantic
pyyaml
requests
apscheduler
sqlalchemy